+++ src/ua/data_types/browse_result.rs
@@ -11,7 +11,7 @@ impl BrowseResult {
     #[must_use]
     pub fn references(&self) -> Option<ua::Array<ua::ReferenceDescription>> {
         // TODO: Adjust signature to return non-owned value instead.
-        ua::Array::from_raw_parts(self.0.references, self.0.referencesSize)
+        ua::Array::from_raw_parts(self.0.referencesSize, self.0.references)
     }
 
     /// Gets continuation point.